python/DBscript/requirements.txt
python-dotenv==1.2.1
mysql-connector-python==9.5.0
pandas==2.3.3
openpyxl==3.1.5
plotly
